Rename all aux.* files to make Windows happy
