+++ src/actions/index.ts
@@ -5,7 +5,7 @@ import { GetOtherActions as createOtherActions } from './other.js'
 import { createBusActions as createBusActions } from './bus.js'
 import { InstanceBaseExt } from '../types.js'
 import { WingConfig } from '../config.js'
-import { createAuxActions } from './aux.js'
+import { createAuxActions } from './auxes.js'
 import { createMainActions } from './main.js'
 import { createUsbPlayerActions } from './usbplayer.js'
 import { createCardsActions } from './cards.js'
