+++ src/commands/auxes.ts
